Optimizing

+++ cosmictiger/complex.hpp
@@ -135,7 +135,7 @@ inline void swap(complex<T>& a, complex<T>& b) {
 
 
 
-using cmplx = complex<double>;
+using cmplx = complex<float>;
 
 
 inline cmplx expc(cmplx z) {
